# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Liste tous les utilisateurs",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "2bb790bde530c0310dc1a0b47a7291a3",
                "requestBody": {
                    "description": "Données de l'utilisateur à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password",
                                    "accept_policy"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "john_doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securePassword123"
                                    },
                                    "accept_policy": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Récupérer un utilisateur par ID",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données utilisateur à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/users/{id}/profil": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Récupérer le profil d'un utilisateur par ID",
                "operationId": "getUserProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour le profil d'un utilisateur",
                "operationId": "updateUserProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du profil à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur ou profil non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "properties": {
                    "firstname": {
                        "type": "string",
                        "example": ""
                    },
                    "lastname": {
                        "type": "string",
                        "example": ""
                    },
                    "location": {
                        "type": "string",
                        "example": ""
                    },
                    "interests": {
                        "type": "string",
                        "example": ""
                    },
                    "bio": {
                        "type": "string",
                        "example": ""
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": ""
                    },
                    "email": {
                        "type": "string",
                        "example": ""
                    },
                    "password": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            }
        }
    },
...
}